chapter_j_001/src/main/java/ru/job4j/generic/RoleStore.java
package ru.job4j.generic;

public class RoleStore implements Store<UserRole> {

    private final Store<UserRole> store = new MemStore<>();

    @Override
    public void add(UserRole model) {
        store.add(model);
    }

    @Override
    public boolean replace(String id, UserRole model) {
        return store.replace(id, model);
    }

    @Override
    public boolean delete(String id) {
        return store.delete(id);
    }

    @Override
    public UserRole findById(String id) {
        return store.findById(id);
    }

}
